--- .travis.yml
@@ -1,38 +1,0 @@
-
-language: cpp
-
-env:
-  global:
-    ExternalData_OBJECT_STORES=${HOME}/.ExternalData
-    KRETZ_SRC_DIR=${TRAVIS_BUILD_DIR}
-    KRETZ_BUILD_DIR=${KRETZ_SRC_DIR}/bld
-    CMAKE_DOWNLOAD_FILE=cmake-3.6.0-Linux-x86_64.sh
-
-cache:
-  directories:
-    - ${ExternalData_OBJECT_STORES}
-
-matrix:
-  include:
-    - os: linux
-      dist: bionic 
-      sudo: required
-
-before_script:
-  - sudo apt-get remove cmake && sudo apt-get install swig lua5.1 doxygen graphviz libboost-program-options-dev
-  - cd ${ExternalData_OBJECT_STORES} && if [[ ! -e ${CMAKE_DOWNLOAD_FILE} ]]; then curl -sSO https://cmake.org/files/v3.6/${CMAKE_DOWNLOAD_FILE}; fi
-  - echo "y\n" | sudo bash "${ExternalData_OBJECT_STORES}/${CMAKE_DOWNLOAD_FILE}" --prefix=/usr/local --exclude-subdir
-  - cmake --version
-  - mkdir -p ${KRETZ_BUILD_DIR}
-  - git clone https://github.com/InsightSoftwareConsortium/ITK.git
-  - mkdir ITK-build; cd ITK-build
-  - cmake ../ITK -DBUILD_EXAMPLES=OFF
-  - make -j 10
-  - sudo make install
-
-script:
-  - cd ${KRETZ_BUILD_DIR}
-  - cmake "${KRETZ_SRC_DIR}" 
-  - make doc 
-  - make  
-  - make test
