+++ include/itkCartesianToToroidalTransform.h
@@ -1,7 +1,22 @@
-
-
-#ifndef __itkCartesianToToroidalTransform_h
-#define __itkCartesianToToroidalTransform_h
+/*=========================================================================
+ *
+ *  Copyright Insight Software Consortium
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0.txt
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ *=========================================================================*/
+#ifndef itkCartesianToToroidalTransform_h
+#define itkCartesianToToroidalTransform_h
 
 
 #include <iostream>
@@ -17,14 +32,17 @@
 namespace itk
 {
 
-/** \brief Toroidal transformation of a vector space (e.g. space coordinates).
+/**
+ * \class CartesianToToroidalTransform
+ *
+ * \brief Toroidal transformation of a vector space (e.g. space coordinates).
  *
  * Transforms three coordinates form toroidal space <alpha,radius> to cartesian
  * coordinates. These are used in trnasfomring the output of 3D ultrasound volumes 
  *
- * \f[			x = r \times sin( \theta ) \f]
- * \f[			y = -sin( \phi ) \times  (r cos( \theta ) -d) \f]
- * \f[			z = d \times (1-cos( \phi ) ) + r \times cos( \theta ) \times cos( \phi ) \f]
+ * \f[      x = r \times sin( \theta ) \f]
+ * \f[      y = -sin( \phi ) \times  (r cos( \theta ) -d) \f]
+ * \f[      z = d \times (1-cos( \phi ) ) + r \times cos( \theta ) \times cos( \phi ) \f]
  *
  *
  * where;
@@ -41,6 +59,7 @@ namespace itk
  * \author Pádraig Looney, University of Oxford.
  *
  * \ingroup Transforms
+ * \ingroup IOKretz
  */
 template <
         class TScalarType=double,          // Data type for scalars (float or double)
@@ -50,11 +69,11 @@ class ITK_EXPORT CartesianToToroidalTransform : public Transform< TScalarType, N
 public:
 
     /** Standard class typedefs. */
-    typedef CartesianToToroidalTransform Self;
+    typedef CartesianToToroidalTransform                       Self;
     typedef Transform< TScalarType, NDimensions, NDimensions > Superclass;
-    typedef SmartPointer<Self>        Pointer;
-    typedef SmartPointer<const Self>  ConstPointer;
-    typedef std::vector<std::pair<double, double> > TableType;
+    typedef SmartPointer<Self>                                 Pointer;
+    typedef SmartPointer<const Self>                           ConstPointer;
+    typedef std::vector<std::pair<double, double> >            TableType;
 
     /** Variables specific to geometry*/
     double m_BModeRadius;
@@ -109,7 +128,7 @@ public:
     OutputPointType;
 
     typedef typename Superclass::FixedParametersType FixedParametersType;
-    typedef typename Superclass::ParametersType ParametersType;
+    typedef typename Superclass::ParametersType      ParametersType;
 
     /** Method to transform a point.
      * This method transforms first two dimensions of a point from polar
