+++ include/itkToroidalToCartesianTransform.h
@@ -1,30 +1,51 @@
-
-
-#ifndef __itkToroidalToCartesianTransform_h
-#define __itkToroidalToCartesianTransform_h
+/*=========================================================================
+ *
+ *  Copyright Insight Software Consortium
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0.txt
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ *=========================================================================*/
+#ifndef itkToroidalToCartesianTransform_h
+#define itkToroidalToCartesianTransform_h
 
 
 #include <iostream>
+#include "itkMesh.h"
 #include "itkTransform.h"
 #include "itkExceptionObject.h"
 #include "itkMatrix.h"
+#include <itkBoundingBox.h>
 #include <vector>
 #include <itkArray.h>
-#include <math.h>
+#include <cmath>
 #include <limits>
 
 
 namespace itk
 {
 
-/** \brief Toroidal transformation of a vector space (e.g. space coordinates).
+/**
+ *
+ * \class ToroidalToCartesianTransform
+ *
+ * \brief Toroidal transformation of a vector space (e.g. space coordinates).
  *
  * Transforms three coordinates form toroidal space <alpha,radius> to cartesian
  * coordinates. These are used in trnasfomring the output of 3D ultrasound volumes 
  *
- * \f[			r = \sqrt{ x^2 + (d - \frac{y}{sin( \phi )} ) } \f]
- * \f[			\phi = -tan^{-1}( \frac{y}{z-d} ) \f]
- * \f[			\theta = sin^{-1}( \frac{x}{b} ) \f]
+ * \f[      r = \sqrt{ x^2 + (d - \frac{y}{sin( \phi )} ) } \f]
+ * \f[      \phi = -tan^{-1}( \frac{y}{z-d} ) \f]
+ * \f[      \theta = sin^{-1}( \frac{x}{b} ) \f]
  *
  *
  * where;
@@ -43,20 +64,21 @@ namespace itk
  * \author Pádraig Looney, University of Oxford.
  *
  * \ingroup Transforms
+ * \ingroup IOKretz
  */
 template <
         class TScalarType=double,          // Data type for scalars (float or double)
         unsigned int NDimensions=3>        // Number of dimensions
-class ITK_EXPORT ToroidalToCartesianTransform : public Transform< TScalarType, NDimensions, NDimensions >
+class ITK_TEMPLATE_EXPORT ToroidalToCartesianTransform : public Transform< TScalarType, NDimensions, NDimensions >
 {
 public:
 
     /** Standard class typedefs. */
-    typedef ToroidalToCartesianTransform Self;
+    typedef ToroidalToCartesianTransform                       Self;
     typedef Transform< TScalarType, NDimensions, NDimensions > Superclass;
-    typedef SmartPointer<Self>        Pointer;
-    typedef SmartPointer<const Self>  ConstPointer;
-    typedef std::vector<std::pair<double, double> > TableType;
+    typedef SmartPointer<Self>                                 Pointer;
+    typedef SmartPointer<const Self>                           ConstPointer;
+    typedef std::vector<std::pair<double, double> >            TableType;
 
     /** Variables specific to geometry*/
     double m_BModeRadius;
@@ -111,7 +133,7 @@ public:
     OutputPointType;
 
     typedef typename Superclass::FixedParametersType FixedParametersType;
-    typedef typename Superclass::ParametersType ParametersType;
+    typedef typename Superclass::ParametersType      ParametersType;
 
     /** Method to transform a point.
    * This method transforms first two dimensions of a point from polar
@@ -167,6 +189,17 @@ public:
     void SetFixedParameters(const FixedParametersType &) override {
 
     }
+
+    typedef Mesh< double, 3 >                         MeshType;
+    typedef typename MeshType::BoundingBoxType        BoundingBoxType;
+    typedef typename BoundingBoxType::BoundsArrayType BoundsArrayType;
+    /*
+     * We need to find the bounds of the toroidal volume in cartesian coordinates
+     * We cannot assume the transformed toroidal boundary points form the cartesian bounds
+     * Hence we need to transform each point and find the bounds of the computed mesh
+     */
+    template< typename TImage >
+    static BoundsArrayType ComputeBounds(const TImage * image, const Self * t2c);
 
 protected:
     ToroidalToCartesianTransform();
