+++ test/itkToroidalTransformTest.cxx
@@ -1,3 +1,20 @@
+/*=========================================================================
+ *
+ *  Copyright Insight Software Consortium
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0.txt
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ *=========================================================================*/
 #include <iostream>
 
 #include "itkToroidalToCartesianTransform.h"
@@ -8,17 +25,114 @@
 #include <cmath>
 
 #include "itkImage.h"
-#include "vnl/vnl_math.h"
 
 #include <itkLabelOverlapMeasuresImageFilter.h>
 
 const unsigned int ParametricDimension = 3;
 
-typedef itk::Image<unsigned char, ParametricDimension> ImageType;
+typedef itk::Image<unsigned char, ParametricDimension>  ImageType;
 typedef itk::LabelOverlapMeasuresImageFilter<ImageType> LabelOverlapType;
 
-std::vector<double> angles1 = {-0.6166550000000001, -0.6097650000000001, -0.602875, -0.5959850000000001, -0.589095, -0.5822050000000001, -0.575315, -0.5684250000000002, -0.5615350000000001, -0.554645, -0.547755, -0.5408650000000002, -0.5339750000000001, -0.527085, -0.520195, -0.5133050000000001, -0.5064150000000001, -0.499525, -0.49263500000000016, -0.4857450000000001, -0.47885500000000003, -0.4719650000000002, -0.4650750000000001, -0.45818500000000006, -0.451295, -0.44440500000000016, -0.4375150000000001, -0.43062500000000004, -0.423735, -0.41684500000000013, -0.40995500000000007, -0.403065, -0.39617500000000017, -0.3892850000000001, -0.38239500000000004, -0.3755050000000002, -0.36861500000000014, -0.3617250000000001, -0.354835, -0.34794499999999995, -0.3410550000000001, -0.33416500000000005, -0.327275, -0.32038500000000014, -0.3134950000000001, -0.306605, -0.2997150000000002, -0.2928250000000001, -0.28593500000000005, -0.2790450000000002, -0.27215500000000015, -0.2652650000000001, -0.258375, -0.25148499999999996, -0.24459500000000012, -0.23770500000000006, -0.230815, -0.22392500000000015, -0.2170350000000001, -0.21014500000000003, -0.20325500000000019, -0.19636500000000012, -0.18947500000000006, -0.18258500000000022, -0.17569500000000016, -0.1688050000000001, -0.16191500000000003, -0.15502499999999997, -0.14813500000000013, -0.14124500000000006, -0.134355, -0.12746500000000016, -0.1205750000000001, -0.11368500000000004, -0.1067950000000002, -0.09990500000000013, -0.09301500000000007, -0.08612500000000023, -0.07923499999999994, -0.0723450000000001, -0.06545500000000026, -0.05856499999999998, -0.05167500000000014, -0.044785000000000075, -0.03789500000000001, -0.03100500000000017, -0.02411500000000011, -0.017225000000000046, -0.010335000000000205, -0.0034450000000001424, 0.0034449999999999203, 0.010334999999999761, 0.017225000000000046, 0.024114999999999887, 0.031004999999999727, 0.03789500000000001, 0.04478499999999985, 0.051674999999999915, 0.05856499999999998, 0.06545499999999982, 0.07234499999999988, 0.07923499999999994, 0.08612499999999979, 0.09301499999999985, 0.09990499999999991, 0.10679499999999975, 0.11368500000000004, 0.12057499999999988, 0.12746499999999972, 0.134355, 0.14124499999999984, 0.1481349999999999, 0.15502499999999997, 0.1619149999999998, 0.16880499999999987, 0.17569499999999993, 0.18258499999999978, 0.18947499999999984, 0.1963649999999999, 0.20325499999999974, 0.21014500000000003, 0.21703499999999987, 0.2239249999999997, 0.230815, 0.23770499999999983, 0.2445949999999999, 0.25148499999999996, 0.2583749999999998, 0.26526499999999986, 0.2721549999999999, 0.27904499999999977, 0.28593499999999983, 0.2928249999999999, 0.29971499999999973, 0.306605, 0.31349499999999986, 0.3203849999999997, 0.327275, 0.3341649999999998, 0.3410549999999999, 0.34794499999999995, 0.3548349999999998, 0.36172499999999985, 0.3686149999999999, 0.37550499999999976, 0.3823949999999998, 0.3892849999999999, 0.3961749999999997, 0.403065, 0.40995499999999985, 0.4168449999999997, 0.423735, 0.4306249999999998, 0.43751499999999965, 0.44440499999999994, 0.4512949999999998, 0.45818500000000006, 0.4650749999999999, 0.47196499999999975, 0.47885500000000003, 0.48574499999999987, 0.4926349999999997, 0.499525, 0.5064149999999998, 0.5133049999999997, 0.520195, 0.5270849999999998, 0.5339749999999996, 0.5408649999999999, 0.5477549999999998, 0.554645, 0.5615349999999999, 0.5684249999999997, 0.575315, 0.5822049999999999, 0.5890949999999997, 0.595985, 0.6028749999999998, 0.6097649999999997, 0.616655};
-std::vector<double> angles2 = {-0.7376595572958549, -0.7271963011639988, -0.7167330450321426, -0.7062697889002866, -0.6958065327684305, -0.6853432766365744, -0.6748800205047183, -0.6644167643728622, -0.6539535082410061, -0.64349025210915, -0.6330269959772938, -0.6225637398454378, -0.6121004837135817, -0.6016372275817256, -0.5911739714498695, -0.5807107153180133, -0.5702474591861573, -0.5597842030543012, -0.5493209469224452, -0.5388576907905891, -0.5283944346587328, -0.5179311785268768, -0.5074679223950207, -0.49700466626316464, -0.4865414101313086, -0.4760781539994525, -0.46561489786759624, -0.4551516417357402, -0.44468838560388413, -0.43422512947202807, -0.423761873340172, -0.41329861720831573, -0.4028353610764597, -0.3923721049446036, -0.38190884881274756, -0.3714455926808915, -0.3609823365490352, -0.35051908041717916, -0.3400558242853231, -0.32959256815346705, -0.319129312021611, -0.30866605588975493, -0.2982027997578989, -0.2877395436260426, -0.27727628749418654, -0.2668130313623305, -0.2563497752304744, -0.24588651909861836, -0.23542326296676208, -0.22496000683490602, -0.21449675070304997, -0.2040334945711939, -0.19357023843933785, -0.18310698230748157, -0.17264372617562573, -0.16218047004376945, -0.1517172139119134, -0.14125395778005734, -0.13079070164820106, -0.12032744551634522, -0.10986418938448894, -0.09940093325263288, -0.08893767712077683, -0.07847442098892055, -0.06801116485706471, -0.05754790872520843, -0.04708465259335237, -0.036621396461496314, -0.026158140329640256, -0.015694884197784198, -0.005231628065927918, 0.00523162806592814, 0.015694884197784198, 0.026158140329640256, 0.036621396461496314, 0.047084652593352594, 0.05754790872520843, 0.06801116485706471, 0.07847442098892077, 0.08893767712077683, 0.0994009332526331, 0.10986418938448894, 0.12032744551634522, 0.13079070164820128, 0.14125395778005734, 0.15171721391191362, 0.16218047004376945, 0.17264372617562573, 0.1831069823074818, 0.19357023843933785, 0.2040334945711939, 0.21449675070304997, 0.22496000683490625, 0.23542326296676208, 0.24588651909861836, 0.2563497752304744, 0.2668130313623305, 0.27727628749418654, 0.2877395436260428, 0.2982027997578989, 0.30866605588975493, 0.319129312021611, 0.32959256815346705, 0.3400558242853233, 0.3505190804171794, 0.36098233654903544, 0.3714455926808915, 0.38190884881274756, 0.3923721049446036, 0.4028353610764599, 0.41329861720831595, 0.423761873340172, 0.4342251294720283, 0.44468838560388413, 0.45515164173573996, 0.4656148978675967, 0.4760781539994525, 0.48654141013130836, 0.49700466626316464, 0.5074679223950209, 0.5179311785268768, 0.528394434658733, 0.5388576907905893, 0.5493209469224452, 0.559784203054301, 0.5702474591861573, 0.5807107153180135, 0.5911739714498694, 0.6016372275817257, 0.6121004837135819, 0.6225637398454378, 0.6330269959772941, 0.6434902521091503, 0.6539535082410062, 0.664416764372862, 0.6748800205047183, 0.6853432766365746, 0.6958065327684304, 0.7062697889002867, 0.716733045032143, 0.7271963011639988, 0.7376595572958546};
+std::vector<double> angles1 = {-0.6166550000000001, -0.6097650000000001,
+  -0.602875, -0.5959850000000001, -0.589095, -0.5822050000000001, -0.575315,
+  -0.5684250000000002, -0.5615350000000001, -0.554645, -0.547755,
+  -0.5408650000000002, -0.5339750000000001, -0.527085, -0.520195,
+  -0.5133050000000001, -0.5064150000000001, -0.499525, -0.49263500000000016,
+  -0.4857450000000001, -0.47885500000000003, -0.4719650000000002,
+  -0.4650750000000001, -0.45818500000000006, -0.451295, -0.44440500000000016,
+  -0.4375150000000001, -0.43062500000000004, -0.423735, -0.41684500000000013,
+  -0.40995500000000007, -0.403065, -0.39617500000000017, -0.3892850000000001,
+  -0.38239500000000004, -0.3755050000000002, -0.36861500000000014,
+  -0.3617250000000001, -0.354835, -0.34794499999999995, -0.3410550000000001,
+  -0.33416500000000005, -0.327275, -0.32038500000000014, -0.3134950000000001,
+  -0.306605, -0.2997150000000002, -0.2928250000000001, -0.28593500000000005,
+  -0.2790450000000002, -0.27215500000000015, -0.2652650000000001, -0.258375,
+  -0.25148499999999996, -0.24459500000000012, -0.23770500000000006, -0.230815,
+  -0.22392500000000015, -0.2170350000000001, -0.21014500000000003,
+  -0.20325500000000019, -0.19636500000000012, -0.18947500000000006,
+  -0.18258500000000022, -0.17569500000000016, -0.1688050000000001,
+  -0.16191500000000003, -0.15502499999999997, -0.14813500000000013,
+  -0.14124500000000006, -0.134355, -0.12746500000000016, -0.1205750000000001,
+  -0.11368500000000004, -0.1067950000000002, -0.09990500000000013,
+  -0.09301500000000007, -0.08612500000000023, -0.07923499999999994,
+  -0.0723450000000001, -0.06545500000000026, -0.05856499999999998,
+  -0.05167500000000014, -0.044785000000000075, -0.03789500000000001,
+  -0.03100500000000017, -0.02411500000000011, -0.017225000000000046,
+  -0.010335000000000205, -0.0034450000000001424, 0.0034449999999999203,
+  0.010334999999999761, 0.017225000000000046, 0.024114999999999887,
+  0.031004999999999727, 0.03789500000000001, 0.04478499999999985,
+  0.051674999999999915, 0.05856499999999998, 0.06545499999999982,
+  0.07234499999999988, 0.07923499999999994, 0.08612499999999979,
+  0.09301499999999985, 0.09990499999999991, 0.10679499999999975,
+  0.11368500000000004, 0.12057499999999988, 0.12746499999999972, 0.134355,
+  0.14124499999999984, 0.1481349999999999, 0.15502499999999997,
+  0.1619149999999998, 0.16880499999999987, 0.17569499999999993,
+  0.18258499999999978, 0.18947499999999984, 0.1963649999999999,
+  0.20325499999999974, 0.21014500000000003, 0.21703499999999987,
+  0.2239249999999997, 0.230815, 0.23770499999999983, 0.2445949999999999,
+  0.25148499999999996, 0.2583749999999998, 0.26526499999999986,
+  0.2721549999999999, 0.27904499999999977, 0.28593499999999983,
+  0.2928249999999999, 0.29971499999999973, 0.306605, 0.31349499999999986,
+  0.3203849999999997, 0.327275, 0.3341649999999998, 0.3410549999999999,
+  0.34794499999999995, 0.3548349999999998, 0.36172499999999985,
+  0.3686149999999999, 0.37550499999999976, 0.3823949999999998,
+  0.3892849999999999, 0.3961749999999997, 0.403065, 0.40995499999999985,
+  0.4168449999999997, 0.423735, 0.4306249999999998, 0.43751499999999965,
+  0.44440499999999994, 0.4512949999999998, 0.45818500000000006,
+  0.4650749999999999, 0.47196499999999975, 0.47885500000000003,
+  0.48574499999999987, 0.4926349999999997, 0.499525, 0.5064149999999998,
+  0.5133049999999997, 0.520195, 0.5270849999999998, 0.5339749999999996,
+  0.5408649999999999, 0.5477549999999998, 0.554645, 0.5615349999999999,
+  0.5684249999999997, 0.575315, 0.5822049999999999, 0.5890949999999997,
+  0.595985, 0.6028749999999998, 0.6097649999999997, 0.616655};
+std::vector<double> angles2 = {-0.7376595572958549, -0.7271963011639988,
+  -0.7167330450321426, -0.7062697889002866, -0.6958065327684305,
+  -0.6853432766365744, -0.6748800205047183, -0.6644167643728622,
+  -0.6539535082410061, -0.64349025210915, -0.6330269959772938,
+  -0.6225637398454378, -0.6121004837135817, -0.6016372275817256,
+  -0.5911739714498695, -0.5807107153180133, -0.5702474591861573,
+  -0.5597842030543012, -0.5493209469224452, -0.5388576907905891,
+  -0.5283944346587328, -0.5179311785268768, -0.5074679223950207,
+  -0.49700466626316464, -0.4865414101313086, -0.4760781539994525,
+  -0.46561489786759624, -0.4551516417357402, -0.44468838560388413,
+  -0.43422512947202807, -0.423761873340172, -0.41329861720831573,
+  -0.4028353610764597, -0.3923721049446036, -0.38190884881274756,
+  -0.3714455926808915, -0.3609823365490352, -0.35051908041717916,
+  -0.3400558242853231, -0.32959256815346705, -0.319129312021611,
+  -0.30866605588975493, -0.2982027997578989, -0.2877395436260426,
+  -0.27727628749418654, -0.2668130313623305, -0.2563497752304744,
+  -0.24588651909861836, -0.23542326296676208, -0.22496000683490602,
+  -0.21449675070304997, -0.2040334945711939, -0.19357023843933785,
+  -0.18310698230748157, -0.17264372617562573, -0.16218047004376945,
+  -0.1517172139119134, -0.14125395778005734, -0.13079070164820106,
+  -0.12032744551634522, -0.10986418938448894, -0.09940093325263288,
+  -0.08893767712077683, -0.07847442098892055, -0.06801116485706471,
+  -0.05754790872520843, -0.04708465259335237, -0.036621396461496314,
+  -0.026158140329640256, -0.015694884197784198, -0.005231628065927918,
+  0.00523162806592814, 0.015694884197784198, 0.026158140329640256,
+  0.036621396461496314, 0.047084652593352594, 0.05754790872520843,
+  0.06801116485706471, 0.07847442098892077, 0.08893767712077683,
+  0.0994009332526331, 0.10986418938448894, 0.12032744551634522,
+  0.13079070164820128, 0.14125395778005734, 0.15171721391191362,
+  0.16218047004376945, 0.17264372617562573, 0.1831069823074818,
+  0.19357023843933785, 0.2040334945711939, 0.21449675070304997,
+  0.22496000683490625, 0.23542326296676208, 0.24588651909861836,
+  0.2563497752304744, 0.2668130313623305, 0.27727628749418654,
+  0.2877395436260428, 0.2982027997578989, 0.30866605588975493,
+  0.319129312021611, 0.32959256815346705, 0.3400558242853233,
+  0.3505190804171794, 0.36098233654903544, 0.3714455926808915,
+  0.38190884881274756, 0.3923721049446036, 0.4028353610764599,
+  0.41329861720831595, 0.423761873340172, 0.4342251294720283,
+  0.44468838560388413, 0.45515164173573996, 0.4656148978675967,
+  0.4760781539994525, 0.48654141013130836, 0.49700466626316464,
+  0.5074679223950209, 0.5179311785268768, 0.528394434658733,
+  0.5388576907905893, 0.5493209469224452, 0.559784203054301,
+  0.5702474591861573, 0.5807107153180135, 0.5911739714498694,
+  0.6016372275817257, 0.6121004837135819, 0.6225637398454378,
+  0.6330269959772941, 0.6434902521091503, 0.6539535082410062,
+  0.664416764372862, 0.6748800205047183, 0.6853432766365746,
+  0.6958065327684304, 0.7062697889002867, 0.716733045032143,
+  0.7271963011639988, 0.7376595572958546};
 
 std::vector<std::pair<double,double>> createTable(std::vector<double> angles){
 
@@ -109,9 +223,15 @@ ImageType::Pointer createImage(ImageType::SizeType size)
     return returnImage;
 }
 
-int main(int ,char *[] )
+int itkToroidalTransformTest(int argc,char *argv[] )
 {
-    const unsigned int Dimension = 3;
+
+    if (argc < 2 )
+      {
+      std::cerr << "Usage: " << argv[0] << " <testImageCartesian>" << std::endl;
+      return EXIT_FAILURE;
+      }
+    constexpr unsigned int Dimension = 3;
 
     //Test is itkToroidalToCartesian after itkCartesianToToroidal identity
 
@@ -120,7 +240,7 @@ int main(int ,char *[] )
 
     typedef itk::ToroidalToCartesianTransform<double,Dimension> T2CTransformType;
     typedef itk::CartesianToToroidalTransform<double,Dimension> C2TTransformType;
-    typedef itk::ImageFileWriter<ImageType> ImageWriterType;
+    typedef itk::ImageFileWriter<ImageType>                     ImageWriterType;
 
     T2CTransformType::Pointer t2c = T2CTransformType::New();
     C2TTransformType::Pointer c2t = C2TTransformType::New();
@@ -128,21 +248,21 @@ int main(int ,char *[] )
     ImageType::DirectionType direction;
     direction.SetIdentity();
 
-    double rBstart=41.599998474121094;
-    double rD=-19.170000076293945;
+    double RadiusBStart=41.599998474121094;
+    double RadiusD=-19.170000076293945;
     double rResol =0.20533333718776703;
 
     ImageType::SpacingType cartesianSpacing;
     cartesianSpacing.Fill(0.6);
 
-    c2t->SetBModeRadius(rD);
-    c2t->SetSweepRadius(rBstart);
+    c2t->SetBModeRadius(RadiusD);
+    c2t->SetSweepRadius(RadiusBStart);
     c2t->SetResolution(rResol);
     c2t->SetTableTheta(TableAngles1);
     c2t->SetTablePhi(TableAngles2);
 
-    t2c->SetBModeRadius(rD);
-    t2c->SetSweepRadius(rBstart);
+    t2c->SetBModeRadius(RadiusD);
+    t2c->SetSweepRadius(RadiusBStart);
     t2c->SetResolution(rResol);
     t2c->SetTableTheta(TableAngles1);
     t2c->SetTablePhi(TableAngles2);
